+++ src/routes/paths.ts
@@ -11,4 +11,7 @@ export const ROUTES = {
   report: {
     keyword: '/report/keyword-letter',
   },
+  setting: {
+    pwReset: '/setting/pw-reset',
+  },
 } as const;
